refactor: reduce GUI header complexity
