+++ packages/gui-web/src/AppWorkspace.tsx
@@ -1,5 +1,5 @@
 /* jshint esversion: 11 */
-import { type ReactElement, type ReactNode, useCallback, useEffect, useRef, useState } from "react";
+import { type ReactElement, type ReactNode, useEffect, useState } from "react";
 import { FiAlertTriangle, FiBell, FiCheckCircle, FiChevronLeft, FiChevronRight, FiCommand, FiFileText, FiGlobe, FiHash, FiHelpCircle, FiInfo, FiLogOut, FiMessageSquare, FiPaperclip, FiSearch, FiSettings, FiShield, FiTerminal, FiTool, FiUser } from "react-icons/fi";
 import type { GuiFileRootId, GuiNotificationSummary, GuiStatusData } from "../../gui-shared/src";
 import { SurfaceGlyph } from "./AppNavigation";
@@ -11,6 +11,7 @@ import { FileExplorerSurface } from "./FileExplorerSurface";
 import { AppsSurface, EditableInventorySurface, InstallationSurface } from "./InventorySurfaces";
 import { AiProvidersSurface, LocalReposSurface, LockedVaultGate, OverviewSurface, PlannedSurface, ProjectsSurface, SecuritySurface, VaultSurface } from "./StatusSurfaces";
 import { isVaultSurfaceLocked, vaultCollectionForSurface } from "./VaultBadges";
+import { applyCommandPaletteSelection, useHeaderMenuState } from "./workspace-header-state";
 
 const communityLinks = {
   github: "https://github.com/marcusquinn/aidevops",
@@ -64,32 +65,12 @@ function WorkspaceHeader({ activeItem, activeSectionLabel, activeSurface, canGoB
   setShellMode: (mode: ShellMode) => void;
   status: GuiStatusData;
 }): ReactElement {
-  const [assistantOpen, setAssistantOpen] = useState(false);
   const [commandInitialQuery, setCommandInitialQuery] = useState("");
   const [commandOpen, setCommandOpen] = useState(false);
-  const [notificationsOpen, setNotificationsOpen] = useState(false);
-  const [profileOpen, setProfileOpen] = useState(false);
-  const headerActionsRef = useRef<HTMLDivElement | null>(null);
-  const closeHeaderTimerRef = useRef<number | null>(null);
+  const headerMenus = useHeaderMenuState();
   const userInitials = status.machine.initials || "AI";
   const userName = status.machine.username || "Local user";
   const activeNotifications = status.notifications.filter((notification) => notification.status === "active");
-  const clearScheduledHeaderClose = useCallback(() => {
-    if (closeHeaderTimerRef.current !== null) {
-      window.clearTimeout(closeHeaderTimerRef.current);
-      closeHeaderTimerRef.current = null;
-    }
-  }, []);
-  const closeHeaderMenus = useCallback(() => {
-    clearScheduledHeaderClose();
-    setAssistantOpen(false);
-    setNotificationsOpen(false);
-    setProfileOpen(false);
-  }, [clearScheduledHeaderClose]);
-  const scheduleHeaderMenusClose = useCallback(() => {
-    clearScheduledHeaderClose();
-    closeHeaderTimerRef.current = window.setTimeout(closeHeaderMenus, 2_400);
-  }, [clearScheduledHeaderClose, closeHeaderMenus]);
 
   useEffect(() => {
     const openCommandPalette = (event: KeyboardEvent) => {
@@ -107,47 +88,17 @@ function WorkspaceHeader({ activeItem, activeSectionLabel, activeSurface, canGoB
     return () => window.removeEventListener("keydown", openCommandPalette);
   }, [goBack, goForward]);
 
-  useEffect(() => {
-    const closeOnPointerDown = (event: PointerEvent) => {
-      const target = event.target;
-      if (target instanceof Node && headerActionsRef.current?.contains(target)) {
-        return;
-      }
-
-      closeHeaderMenus();
-    };
-    const closeOnEscape = (event: KeyboardEvent) => {
-      if (event.key === "Escape") {
-        closeHeaderMenus();
-      }
-    };
-
-    document.addEventListener("pointerdown", closeOnPointerDown, true);
-    document.addEventListener("keydown", closeOnEscape);
-    return () => {
-      document.removeEventListener("pointerdown", closeOnPointerDown, true);
-      document.removeEventListener("keydown", closeOnEscape);
-      clearScheduledHeaderClose();
-    };
-  }, [clearScheduledHeaderClose, closeHeaderMenus]);
-
-  const closeMenus = () => {
-    closeHeaderMenus();
-  };
-
-  const openItem = ({ conversationMode, repoPathRef, sessionId, shellMode, surface }: CommandPaletteSelection) => {
-    const repoIndex = status.local_repos.repos.findIndex((repo) => repo.path_ref === repoPathRef);
-    setShellMode(shellMode ?? "devices");
-    if (repoIndex >= 0) {
-      setSelectedLocalRepoIndex(repoIndex);
-    }
-    setSelectedSessionId(sessionId);
-    if (conversationMode) {
-      setConversationMode(conversationMode);
-    }
-    setActiveSurface(surface);
-    closeMenus();
-    setCommandOpen(false);
+  const openItem = (selection: CommandPaletteSelection) => {
+    applyCommandPaletteSelection(selection, {
+      closeCommandPalette: () => setCommandOpen(false),
+      closeHeaderMenus: headerMenus.closeHeaderMenus,
+      setActiveSurface,
+      setConversationMode,
+      setSelectedLocalRepoIndex,
+      setSelectedSessionId,
+      setShellMode,
+      status,
+    });
   };
 
   return (
@@ -170,28 +121,28 @@ function WorkspaceHeader({ activeItem, activeSectionLabel, activeSurface, canGoB
           <strong>{text.searchPlaceholder}</strong>
         </button>
       </div>
-      <div className="header-actions" onPointerEnter={clearScheduledHeaderClose} onPointerLeave={scheduleHeaderMenusClose} ref={headerActionsRef}>
+      <div className="header-actions" onPointerEnter={headerMenus.clearScheduledHeaderClose} onPointerLeave={headerMenus.scheduleHeaderMenusClose} ref={headerMenus.headerActionsRef}>
         <div className="header-action-menu">
-          <button aria-label="Open signposts and help" className="header-icon-button" onClick={() => { closeHeaderMenus(); openItem({ surface: "help" }); }} title="Signposts and help (?)" type="button">
+          <button aria-label="Open signposts and help" className="header-icon-button" onClick={() => { headerMenus.closeHeaderMenus(); openItem({ surface: "help" }); }} title="Signposts and help (?)" type="button">
             <FiHelpCircle aria-hidden="true" />
           </button>
-          <button aria-expanded={notificationsOpen} aria-label="Open notifications" className="header-icon-button" onClick={() => { clearScheduledHeaderClose(); setNotificationsOpen((current) => !current); setProfileOpen(false); setAssistantOpen(false); }} type="button">
+          <button aria-expanded={headerMenus.notificationsOpen} aria-label="Open notifications" className="header-icon-button" onClick={headerMenus.toggleNotifications} type="button">
             <FiBell aria-hidden="true" />
             {activeNotifications.length > 0 ? <span className="notification-dot" aria-hidden="true" /> : null}
           </button>
-          {notificationsOpen ? <NotificationsMenu notifications={status.notifications} openSurface={(surface) => openItem({ surface })} /> : null}
+          {headerMenus.notificationsOpen ? <NotificationsMenu notifications={status.notifications} openSurface={(surface) => openItem({ surface })} /> : null}
         </div>
-        <button aria-pressed={assistantOpen} aria-label="Toggle AI Assistant" className={assistantOpen ? "header-icon-button active" : "header-icon-button"} onClick={() => { clearScheduledHeaderClose(); setAssistantOpen((current) => !current); setNotificationsOpen(false); setProfileOpen(false); }} title="AI sessions (_)" type="button">
+        <button aria-pressed={headerMenus.assistantOpen} aria-label="Toggle AI Assistant" className={headerMenus.assistantOpen ? "header-icon-button active" : "header-icon-button"} onClick={headerMenus.toggleAssistant} title="AI sessions (_)" type="button">
           <FiTerminal aria-hidden="true" />
         </button>
         <div className="header-action-menu">
-          <button aria-expanded={profileOpen} aria-label={`Open profile menu for ${userName}`} className="profile-avatar-button" onClick={() => { clearScheduledHeaderClose(); setProfileOpen((current) => !current); setNotificationsOpen(false); setAssistantOpen(false); }} title={userName} type="button">
+          <button aria-expanded={headerMenus.profileOpen} aria-label={`Open profile menu for ${userName}`} className="profile-avatar-button" onClick={headerMenus.toggleProfile} title={userName} type="button">
             <span>{userInitials}</span>
           </button>
-          {profileOpen ? <ProfileMenu openSurface={(surface) => openItem({ surface })} userName={userName} /> : null}
+          {headerMenus.profileOpen ? <ProfileMenu openSurface={(surface) => openItem({ surface })} userName={userName} /> : null}
         </div>
       </div>
-      {assistantOpen ? <AssistantPanel activeSurface={activeSurface} userName={userName} /> : null}
+      {headerMenus.assistantOpen ? <AssistantPanel activeSurface={activeSurface} userName={userName} /> : null}
       {commandOpen ? <CommandPalette activeSurface={activeSurface} close={() => setCommandOpen(false)} initialQuery={commandInitialQuery} openItem={openItem} status={status} /> : null}
     </header>
   );
